refactor(log): move LogErrorRepository to folder
