+++ src/main/decorators/log-contoller-decorator.ts
@@ -1,4 +1,4 @@
-import { LogErrorRepository } from '../../data/protocols/db/log-error-repository'
+import { LogErrorRepository } from '../../data/protocols/db/log/log-error-repository'
 import { Controller, HttpRequest, HttpRespose } from '../../presentation/protocols'
 
 export class LogControllerDecorator implements Controller {
